MYR-2059: sepereate code in file function for pallet tipping
