+++ pallets/tipping/src/functions/mod.rs
@@ -1,0 +1,4 @@
+mod api;
+mod helper;
+mod tips_balance;
+mod unsigned_call;
